code-examples/realtime-data-connection/python/raw-websockets/requirements.txt
websockets>=11.0
python-dotenv>=1.0.0